Downloader
The archive rules are kept in a simple format, each rule is only the top section of the url.  Each url can have multiple rules however each rule also has it's own specific rules page.  Thus each page is scraped for what url the specific rule is located on.  From the top of the span to the second h1 ref is considered to be the rule of the webpage.
max_rule = 1975
base_url = f'https://2e.aonprd.com/Rules.aspx?ID={0 to max_rule}'